Potential crash prevention by disabling user interaction on calendar collection view during scope transition
When scope is being transitioned from week to month or month to week
and there is another scroll view scrolling in the view, there is a case
where the calendar will crash the app. Additionally, it doesn't really
make sense that the calendar can be scrolled during the transition. Fix
involves disabling user interaction when the transition begins and is
re-enabled after the transition ends.